backend/requirements.txt
fastapi
uvicorn
crewai
google-generativeai
python-dotenv
tavily-python
pydantic